+++ src/model/dao/TableDao.java
@@ -6,6 +6,6 @@ public interface TableDao {
     void insert(Table reservation);
     void delete(Table reservation);
     void update(Table reservation);
-    Table findById(Long id);
+    Table findById(int id);
     Table findAll();
 }
